ref: split to core files
